+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 <a name="0.0.1"></a>
-## 0.0.1 (2017-12-07)
+## 0.0.1 (2018-01-02)
 
 
 ### Bug Fixes
@@ -9,6 +9,21 @@
 
 ### Features
 
+* add Animated ([fdc3dc4](https://github.com/yanyixin/ym-react-native/commit/fdc3dc4))
 * add changelog script ([d1ba6a7](https://github.com/yanyixin/ym-react-native/commit/d1ba6a7))
+* **FlatList:** add Loading ([40866b0](https://github.com/yanyixin/ym-react-native/commit/40866b0))
 * add commitizen ([12a9ac1](https://github.com/yanyixin/ym-react-native/commit/12a9ac1))
+* **page:** use easyMock ([9c88e9c](https://github.com/yanyixin/ym-react-native/commit/9c88e9c))
+* add default font family ([698a359](https://github.com/yanyixin/ym-react-native/commit/698a359))
+* **android:** 增加 android 内置字体 ([7847c0f](https://github.com/yanyixin/ym-react-native/commit/7847c0f))
+* **animated:** add Animated.View ([71eebb8](https://github.com/yanyixin/ym-react-native/commit/71eebb8))
+* **flatList:** add FlatList ([d0d7bc9](https://github.com/yanyixin/ym-react-native/commit/d0d7bc9))
+* **FlatList:** updpate ([9544180](https://github.com/yanyixin/ym-react-native/commit/9544180))
+* **Header:** add Header component ([cca802f](https://github.com/yanyixin/ym-react-native/commit/cca802f))
+* **onEndReached:** onEndReached ([2014810](https://github.com/yanyixin/ym-react-native/commit/2014810))
 * **package:** update ([8eab3f4](https://github.com/yanyixin/ym-react-native/commit/8eab3f4))
+* **WebView:** add WebView ([2c8f1ac](https://github.com/yanyixin/ym-react-native/commit/2c8f1ac))
+* **WebView:** update ([cec557d](https://github.com/yanyixin/ym-react-native/commit/cec557d))
+
+
+